requirements.txt
RealtimeSTT>=0.3.0
faster-whisper>=1.0.0
httpx>=0.27.0
anthropic>=0.34.0
keyboard>=0.13.5
pyperclip>=1.9.0
pystray>=0.19.5
Pillow>=10.4.0
python-dotenv>=1.0.0
pytest>=8.0.0
pyinstaller>=6.0.0
